Fix VSTest naming
