+++ source/Nuke.Common/IconClasses.cs
@@ -28,7 +28,7 @@ using Nuke.Common.Tools.Paket;
 using Nuke.Common.Tools.ReportGenerator;
 using Nuke.Common.Tools.SignTool;
 using Nuke.Common.Tools.TestCloud;
-using Nuke.Common.Tools.VsTest;
+using Nuke.Common.Tools.VSTest;
 using Nuke.Common.Tools.WebConfigTransformRunner;
 using Nuke.Common.Tools.Xunit;
 using Nuke.Common;
@@ -78,7 +78,7 @@ using Nuke.Common.Tools.VSWhere;
 [assembly: IconClass(typeof(SolutionAttribute), "tree7")]
 [assembly: IconClass(typeof(TextTasks), "file-text3")]
 [assembly: IconClass(typeof(TestCloudTasks), "bug2")]
-[assembly: IconClass(typeof(VsTestTasks), "bug2")]
+[assembly: IconClass(typeof(VSTestTasks), "bug2")]
 [assembly: IconClass(typeof(VSWhereTasks), "search")]
 [assembly: IconClass(typeof(WebConfigTransformRunnerTasks), "file-empty2")]
 [assembly: IconClass(typeof(XmlTasks), "file-empty2")]
